Issue #67 Updates implicit conversion error in QuatTest
- Fixes issue in QuatTests class that won't compile in Linux due to incorrect conversion from vec4d to vec4f.

+++ Tests/TrBase/QuatTests.h
@@ -48,7 +48,7 @@ public:
     trBase::Quat mQuat = trBase::Quat(0.635, 0.431, 0.390, 0.508);
     osg::Quat mOsgQuat = osg::Quat(0.635, 0.431, 0.390, 0.508);
     trBase::Vec4d mTestQuatVecD = trBase::Vec4d(0.635, 0.431, 0.390, 0.508);
-    trBase::Vec4f mTestQuatVecF = trBase::Vec4d(0.635f, 0.431f, 0.390f, 0.508f);
+    trBase::Vec4f mTestQuatVecF = trBase::Vec4f(0.635f, 0.431f, 0.390f, 0.508f);
     trBase::Vec3d mTestRadVec = trBase::Vec3d(-0.436, 1.204, 2.094);
     trBase::Vec3d mTestDegVec = trBase::Vec3d(-25.000, 69.000, 120.000);
